dcload: Refactor of syscall init + cleanup of dbgio init.
Before, when dbgio couldn't be set to null, the dcload
handler needed to have stubs and would then copy in the
correct functions when initializing. This was partially wrong
from the start as it installed the functions before testing
if dcload was available.

This is all simpler now, the handler simply has its normal functions
and gets added to the list if the syscall's detected and otherwise
it doesn't.

Unfortunately there was a bit of a snarl. The `fs_dcload_init_console`
function needed to be refactored to allow the handler to be installed
after the scif one, and maintain it's priority. This is because the
detection mechanism used by scif to determine which dcload syscall is
installed was being initialized by that same init_console function.
The solution is more properly separating out the detection of dcload
syscall from the installation of the dbgio handler.
